requirements.txt
setuptools>=68.0.0
wheel>=0.41.0
streamlit>=1.28.0
Pillow>=10.0.0
numpy>=1.24.0
paddlepaddle==2.6.0
paddleocr==2.9.0
matplotlib>=3.7.0
pandas>=2.0.0

PyWavelets==1.8.0
opencv-python-headless>=4.8.0